+++ iroh-net/src/hp/stun.rs
@@ -166,7 +166,7 @@ pub mod test {
         let mut m = DerpMap::default();
 
         for (i, addr) in stun.enumerate() {
-            let region_id = i + 1;
+            let region_id = (i + 1) as u16;
             let host = addr.ip();
             let port = addr.port();
